requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiomysql
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
attrs==22.2.0
black
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
isort
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
mysql
pydantic==1.10.6
python-dotenv==1.0.0
pyyaml
SQLAlchemy==2.0.7
typing_extensions==4.5.0
yarl==1.8.2
